# docs/openapi.yaml
...
components:
  schemas:
...
    ComplaintCategorySummary:
      description: 'A complaint category and which departments claim it.


        ``departmentIds`` has more than one entry exactly when 0011''s A2 SLA

        tie-break is load-bearing, which is why it is exposed rather than flattened

        to a single owner.'
      properties:
        departmentIds:
          items:
            type: string
          title: Departmentids
          type: array
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        slaHours:
          anyOf:
          - type: integer
          - type: 'null'
          title: Slahours
        status:
          default: active
          title: Status
          type: string
      required:
      - id
      - name
      title: ComplaintCategorySummary
      type: object
...
    CreateDepartmentRequest:
      description: Create a department, its category claims and its roster in one call.
      properties:
        categories:
          items:
            type: string
          title: Categories
          type: array
        description:
          anyOf:
          - maxLength: 1000
            type: string
          - type: 'null'
          title: Description
        email:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          title: Email
        head:
          anyOf:
          - maxLength: 120
            type: string
          - type: 'null'
          title: Head
        kind:
          anyOf:
          - type: string
          - type: 'null'
          description: service | security
          title: Kind
        name:
          maxLength: 120
          minLength: 1
          title: Name
          type: string
        operatingHours:
          anyOf:
          - $ref: '#/components/schemas/OperatingHours-Input'
          - type: 'null'
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
        slaHours:
          anyOf:
          - exclusiveMinimum: 0.0
            maximum: 8760.0
            type: integer
          - type: 'null'
          title: Slahours
        staff:
          items:
            $ref: '#/components/schemas/StaffMemberInput'
          title: Staff
          type: array
        status:
          anyOf:
          - type: string
          - type: 'null'
          description: Active | Inactive
          title: Status
      required:
      - name
      title: CreateDepartmentRequest
      type: object
...
    DepartmentDetail:
      description: A department plus its roster.
      properties:
        activeComplaintCount:
          default: 0
          title: Activecomplaintcount
          type: integer
        categories:
          items:
            type: string
          title: Categories
          type: array
        categoryIds:
          items:
            type: string
          title: Categoryids
          type: array
        createdAt:
          format: date-time
          title: Createdat
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        head:
          anyOf:
          - type: string
          - type: 'null'
          title: Head
        headStaffId:
          anyOf:
          - type: string
          - type: 'null'
          title: Headstaffid
        id:
          title: Id
          type: string
        kind:
          default: service
          title: Kind
          type: string
        name:
          title: Name
          type: string
        operatingHours:
          $ref: '#/components/schemas/app__domain__department_schemas__OperatingHours'
        overdueComplaintCount:
          default: 0
          title: Overduecomplaintcount
          type: integer
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        resolvedComplaintCount:
          default: 0
          title: Resolvedcomplaintcount
          type: integer
        slaHours:
          anyOf:
          - type: integer
          - type: 'null'
          title: Slahours
        staff:
          items:
            $ref: '#/components/schemas/StaffMember'
          title: Staff
          type: array
        staffCount:
          default: 0
          title: Staffcount
          type: integer
        status:
          default: Active
          title: Status
          type: string
        updatedAt:
          format: date-time
          title: Updatedat
          type: string
      required:
      - id
      - name
      - createdAt
      - updatedAt
      title: DepartmentDetail
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MessageResult:
      description: Acknowledgement for an action with nothing meaningful to return.
      properties:
        message:
          title: Message
          type: string
      required:
      - message
      title: MessageResult
      type: object
...
    Page_ComplaintCategorySummary_:
      properties:
        hasMore:
          title: Hasmore
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/ComplaintCategorySummary'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - pageSize
      - hasMore
      title: Page[ComplaintCategorySummary]
      type: object
...
    Page_DepartmentDetail_:
      properties:
        hasMore:
          title: Hasmore
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/DepartmentDetail'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - pageSize
      - hasMore
      title: Page[DepartmentDetail]
      type: object
...
    ReplaceStaffRequest:
      description: 'The whole roster, as the department form submits it.


        Members not listed are deactivated rather than deleted (A7) -- a complaint''s

        ``assignee`` records staff by name, so a deleted row turns a past assignment

        into an unexplained string.'
      properties:
        staff:
          items:
            $ref: '#/components/schemas/StaffMemberInput'
          title: Staff
          type: array
      title: ReplaceStaffRequest
      type: object
...
    StaffMember:
      description: 'One person on a department''s roster.


        ``membershipId`` is nullable and usually null: C1 established that the

        frontend records only a name and a phone number, so most staff have no

        account. ``name`` is therefore authoritative and the link is the enrichment,

        not the other way round.'
      properties:
        activeAssignmentCount:
          default: 0
          title: Activeassignmentcount
          type: integer
        id:
          title: Id
          type: string
        membershipId:
          anyOf:
          - type: string
          - type: 'null'
          title: Membershipid
        name:
          title: Name
          type: string
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        rank:
          default: member
          title: Rank
          type: string
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        shift:
          anyOf:
          - type: string
          - type: 'null'
          title: Shift
        status:
          default: active
          title: Status
          type: string
      required:
      - id
      - name
      title: StaffMember
      type: object
    StaffMemberInput:
      description: 'A staff member as submitted by the department form.


        ``id`` is optional: the frontend generates a local id for new rows and keeps

        the server id for existing ones, so presence is what distinguishes an update

        from an insert.'
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
        name:
          maxLength: 120
          minLength: 1
          title: Name
          type: string
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
        role:
          anyOf:
          - maxLength: 60
            type: string
          - type: 'null'
          title: Role
        shift:
          anyOf:
          - type: string
          - type: 'null'
          description: Day | Evening | Night
          title: Shift
        status:
          anyOf:
          - type: string
          - type: 'null'
          description: active | inactive
          title: Status
      required:
      - name
      title: StaffMemberInput
      type: object
...
    UpdateDepartmentRequest:
      description: 'Partial update. An omitted field is left alone; an explicit ``null``

        clears it.


        The distinction is real and is preserved all the way to Postgres: the service

        sends only the keys that were set, and the RPC tests ``patch ? ''key''``.

        Sending ``staff`` replaces the whole roster; omitting it leaves it untouched.'
      properties:
        categories:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Categories
        description:
          anyOf:
          - maxLength: 1000
            type: string
          - type: 'null'
          title: Description
        email:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          title: Email
        head:
          anyOf:
          - maxLength: 120
            type: string
          - type: 'null'
          title: Head
        kind:
          anyOf:
          - type: string
          - type: 'null'
          description: service | security
          title: Kind
        name:
          anyOf:
          - maxLength: 120
            minLength: 1
            type: string
          - type: 'null'
          title: Name
        operatingHours:
          anyOf:
          - $ref: '#/components/schemas/OperatingHours-Input'
          - type: 'null'
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
        slaHours:
          anyOf:
          - exclusiveMinimum: 0.0
            maximum: 8760.0
            type: integer
          - type: 'null'
          title: Slahours
        staff:
          anyOf:
          - items:
              $ref: '#/components/schemas/StaffMemberInput'
            type: array
          - type: 'null'
          title: Staff
        status:
          anyOf:
          - type: string
          - type: 'null'
          description: Active | Inactive
          title: Status
      title: UpdateDepartmentRequest
      type: object
...
    UpdateStaffMemberRequest:
      description: Patch one roster entry.
      properties:
        name:
          anyOf:
          - maxLength: 120
            minLength: 1
            type: string
          - type: 'null'
          title: Name
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
        role:
          anyOf:
          - maxLength: 60
            type: string
          - type: 'null'
          title: Role
        shift:
          anyOf:
          - type: string
          - type: 'null'
          description: Day | Evening | Night
          title: Shift
        status:
          anyOf:
          - type: string
          - type: 'null'
          description: active | inactive
          title: Status
      title: UpdateStaffMemberRequest
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  summary: Backend for the HomeBandhu residential community platform.
  title: HomeBandhu API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/complaint-categories:
    get:
      description: 'The category vocabulary behind the department form''s checkbox list.


        ``departmentIds`` carries every department claiming the category. More than

        one entry is the case where the SLA tie-break in migration 0011 decides which

        deadline applies -- exposed rather than hidden, because it is the open

        product question in `DECISIONS_NEEDED.md` A1.'
      operationId: list_categories_api_v1_complaint_categories_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ComplaintCategorySummary_'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: List complaint categories
      tags:
      - departments
...
  /api/v1/departments:
    get:
      description: 'Departments with their staff, category claims and complaint counts.


        ``activeComplaintCount`` and ``overdueComplaintCount`` are computed from the

        complaints actually routed to the department, so they agree with the

        complaints list rather than being re-derived per screen.'
      operationId: list_departments_api_v1_departments_get
      parameters:
      - description: Matches name, description, head, contact email, category names and staff names --
          the same fields the dashboard's search box covers.
        in: query
        name: q
        required: false
        schema:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          description: Matches name, description, head, contact email, category names and staff names
            -- the same fields the dashboard's search box covers.
          title: Q
      - description: Active | Inactive. Omit for both.
        in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Active | Inactive. Omit for both.
          title: Status
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: pageSize
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Pagesize
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_DepartmentDetail_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List departments
      tags:
      - departments
    post:
      description: 'Create a department together with its categories, roster and head.


        All four in one transaction. ``categories`` are names: any that do not exist

        yet are created, because one of the two create screens is a free-text box.


        Returns 409 when the community already has a department with that name.'
      operationId: create_department_api_v1_departments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create a department
      tags:
      - departments
  /api/v1/departments/{department_id}:
    delete:
      description: 'Permanently remove a department, its category claims and its roster.


        **This one really is a delete**, unlike removing a resident. Complaint

        records survive with their department reference cleared, which is what the

        confirmation dialog promises.


        Returns 409 while the department owns any open complaint; the count is taken

        inside the deleting transaction, so a complaint raised mid-check cannot slip

        through. Deactivate instead.'
      operationId: delete_department_api_v1_departments__department_id__delete
      parameters:
      - in: path
        name: department_id
        required: true
        schema:
          title: Department Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResult'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete a department
      tags:
      - departments
    get:
      description: One department with its active roster and each member's open workload.
      operationId: get_department_api_v1_departments__department_id__get
      parameters:
      - in: path
        name: department_id
        required: true
        schema:
          title: Department Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get a department
      tags:
      - departments
    patch:
      description: 'Patch a department. Omitted fields are left unchanged.


        Two fields have collection semantics: sending ``categories`` replaces the

        claim set, and sending ``staff`` replaces the roster. Omitting either leaves

        it alone.


        Setting ``status`` to ``Inactive`` is the deactivation the dashboard offers

        instead of deletion, and is **not** blocked by open complaints -- that guard

        belongs to `DELETE` alone, or an admin with an unresolvable department would

        have no action left.'
      operationId: update_department_api_v1_departments__department_id__patch
      parameters:
      - in: path
        name: department_id
        required: true
        schema:
          title: Department Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update a department
      tags:
      - departments
  /api/v1/departments/{department_id}/staff:
    post:
      description: 'Add one person to a roster.


        A staff member needs no account: name is the only required field, matching

        what the department form actually collects.'
      operationId: add_staff_member_api_v1_departments__department_id__staff_post
      parameters:
      - in: path
        name: department_id
        required: true
        schema:
          title: Department Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffMemberInput'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffMember'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Add a staff member
      tags:
      - departments
    put:
      description: 'Reconcile the whole roster in one transaction.


        Entries carrying an ``id`` are updated in place, entries without one are

        added, and active members the payload omits are **deactivated, not deleted**

        -- a complaint''s ``assignee`` records staff by name, so removing the row

        would turn a past assignment into an unattributable string.


        PUT rather than POST because this replaces a collection: sending the same

        roster twice leaves the same result.'
      operationId: replace_staff_api_v1_departments__department_id__staff_put
      parameters:
      - in: path
        name: department_id
        required: true
        schema:
          title: Department Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceStaffRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/StaffMember'
                title: Response Replace Staff Api V1 Departments  Department Id  Staff Put
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Replace a department's roster
      tags:
      - departments
  /api/v1/departments/{department_id}/staff/{staff_id}:
    delete:
      description: 'Take a member off the active roster.


        **Not a delete.** The row is marked inactive so past assignments recorded

        against their name stay explainable. Removing the head also frees the head

        slot in the same statement.'
      operationId: remove_staff_member_api_v1_departments__department_id__staff__staff_id__delete
      parameters:
      - in: path
        name: department_id
        required: true
        schema:
          title: Department Id
          type: string
      - in: path
        name: staff_id
        required: true
        schema:
          title: Staff Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResult'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Remove a staff member
      tags:
      - departments
    patch:
      description: 'Patch one roster entry.


        Promotion to head is not done here: ``PATCH /departments/{id}`` with a

        ``head`` name is the only path, because promoting one person has to demote

        the incumbent in the same transaction.'
      operationId: update_staff_member_api_v1_departments__department_id__staff__staff_id__patch
      parameters:
      - in: path
        name: department_id
        required: true
        schema:
          title: Department Id
          type: string
      - in: path
        name: staff_id
        required: true
        schema:
          title: Staff Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStaffMemberRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffMember'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update a staff member
      tags:
      - departments
...
servers:
- description: Local development
  url: http://localhost:8000
- description: Deployed
  url: https://{host}
  variables:
    host:
      default: api.homebandhu.example
      description: API host
...